Sweep contributor-clarity traps: callback naming, dead SAML accessor, IP-classifier home, asset casing
Four small, disjoint readability fixes bundled as one sweep (in the style of
#237), re-verified against the code as it stands after the OIDC/SAML flow
extraction (#500/#501), the rate-limit gate move (#503) and the mode-token
typing (#504):

- Renamed the controller action methods OidPost -> OidCallback and
  SamlPost -> SamlCallback for symmetry (OidPost handled a GET, excused only
  by a comment explaining the misleading name). Route attribute literals are
  untouched, so the wire contract does not change. Renamed the SAML flow
  service's Post -> Callback to match, since the OIDC sibling already reads
  CallbackAsync after #501.
- Deleted SamlResponse.GetCustomAttribute (singular) in Saml.cs: a
  whole-solution grep confirms it had no production caller, only the two
  test assertions the issue called out. Retargeted them onto the plural
  GetCustomAttributes (already the only form production code uses) and
  hoisted the repeated "Role" attribute-name literal in SamlLoginService
  into one RoleAttributeName constant.
- Extracted the public/blocked-address classifier (IsBlockedAddress,
  TryExtractEmbeddedIPv4) out of AvatarUrlValidator into a new, neutrally
  named IpAddressClassifier. It still lived only in the avatar-specific file
  even after #503, despite SsoRateLimiter.NormalizeClientKey depending on it
  too - tuning the avatar SSRF policy could silently change which clients
  the login rate limiter exempts. Zero behavior change (the method bodies
  moved verbatim); split the test coverage to match.
- Renamed Views/apiClient.js to Views/ApiClient.js so the on-disk file
  matches the name it was already published and imported under (SSOPlugin's
  GetViews table, linking.html's import) instead of only agreeing
  case-insensitively. Commented the GetPages/GetViews tables as the plugin's
  public URL contract, including why style.css is intentionally published
  under two names.

Also recorded the Oidc-over-Oid naming convention for new OpenID types in
ARCHITECTURE.md, as suggested alongside the four fixes above (no rename of
existing serialized names or routes).

Closes #370
