+++ SSO-Auth/SSO-Auth.csproj
@@ -33,7 +33,7 @@
     <None Remove="Config\config.js" />
     <None Remove="Config\style.css" />
     <None Remove="Config\linking.html" />
-    <None Remove="Views\apiClient.js" />
+    <None Remove="Views\ApiClient.js" />
     <None Remove="Views\jellyfin-apiClient.esm.min.js" />
     <None Remove="Views\emby-restyle.css" />
     <EmbeddedResource Include="Config\configPage.html" />
@@ -41,7 +41,7 @@
     <EmbeddedResource Include="Config\style.css" />
     <EmbeddedResource Include="Config\linking.html" />
     <EmbeddedResource Include="Config\linking.js" />
-    <EmbeddedResource Include="Views\apiClient.js" />
+    <EmbeddedResource Include="Views\ApiClient.js" />
     <EmbeddedResource Include="Views\jellyfin-apiClient.esm.min.js" />
     <EmbeddedResource Include="Views\emby-restyle.css" />
   </ItemGroup>
