+++ SSO-Auth/SSOPlugin.cs
@@ -108,6 +108,20 @@ public class SSOPlugin : BasePlugin<PluginConfiguration>, IHasWebPages
     // overridden pipeline above.
     private void PersistBase(BasePluginConfiguration configuration) => base.UpdateConfiguration(configuration);
 
+    // Both tables below are the plugin's public URL contract (#370): the first element of each Page()
+    // pair is the name a caller (the admin config page, the linking page, SSOViewsController) requests
+    // an asset by, matched case-sensitively (SSOViewsController.GetView); the second is the embedded
+    // resource path suffix, which must match the source file's actual on-disk name and casing under
+    // this project's default (path-derived) embedded-resource naming. The two never have to agree with
+    // each other, but changing either one changes what breaks: renaming the registered name breaks
+    // every caller of that URL (config.js, linking.html, config page markup); renaming/moving the
+    // source file without updating the resource suffix here breaks the embedded-resource lookup at
+    // runtime (a 404, since GetManifestResourceStream is also case-sensitive). Config.style.css is
+    // deliberately published under two different registered names below — "SSO-Auth.css" (GetPages, the
+    // admin config page's own stylesheet load) and "style.css" (GetViews, the public linking page) —
+    // the same resource, two unrelated consumers with independently-chosen URL conventions, not a
+    // casing inconsistency.
+
     /// <summary>
     /// Returns the available internal web pages of this plugin.
     /// </summary>
@@ -132,7 +146,7 @@ public class SSOPlugin : BasePlugin<PluginConfiguration>, IHasWebPages
             Page("style.css", "Config.style.css"),
             Page("linking", "Config.linking.html"),
             Page("linking.js", "Config.linking.js"),
-            Page("ApiClient.js", "Views.apiClient.js"),
+            Page("ApiClient.js", "Views.ApiClient.js"),
             Page("emby-restyle.css", "Views.emby-restyle.css"),
             Page("jellyfin-apiClient.esm.min.js", "Views.jellyfin-apiClient.esm.min.js"),
         };
